backend/pyproject.toml
[project]
name = "etsy-assistant-backend"
version = "0.1.0"
description = "Etsy Assistant web backend - FastAPI on AWS Lambda"
requires-python = ">=3.12"
dependencies = [
    "opencv-python-headless>=4.9",
    "Pillow>=10.0",
    "numpy>=1.26",
    "anthropic>=0.40",
    "httpx>=0.27",
    "fastapi>=0.115",
    "mangum>=0.19",
    "boto3>=1.35",
    "uvicorn>=0.34",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/api"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[dependency-groups]
dev = ["pytest>=8.0", "moto[dynamodb,s3]>=5.0", "httpx>=0.27"]
